rainclaw/websearch/requirements.txt
crawl4ai
fastapi
uvicorn[standard]
gunicorn
aiohttp
pydantic
python-dotenv
markdown
beautifulsoup4